// tlm107-film/src/assets.generated.ts
export const ASSETS: Record<string, Asset> = {
// ...
 "black_front": {
  "slug": "black_front",
  "file": "img/black_front.png",
  "kind": "c",
  "label": "TLM 107 \u00b7 black",
  "w": 1534,
  "h": 3200,
  "ar": 0.4794,
  "lum": 0.268
 },
// ...
 "nickel_front_tall": {
  "slug": "nickel_front_tall",
  "file": "img/nickel_front_tall.png",
  "kind": "c",
  "label": "TLM 107 \u00b7 nickel",
  "w": 1498,
  "h": 3200,
  "ar": 0.4681,
  "lum": 0.651
 },
// ...
 "black_ea4": {
  "slug": "black_ea4",
  "file": "img/black_ea4.png",
  "kind": "c",
  "label": "TLM 107 bk \u00b7 EA 4",
  "w": 2471,
  "h": 2784,
  "ar": 0.8876,
  "lum": 0.261
 },
// ...
};